Renamed to controller

+++ src/MovementController.ino
@@ -80,7 +80,7 @@ DynamicJsonDocument getNetworkScanInfo() {
 }
 
 unsigned long lastMsg = 0;  // Time report control
-int msgPeriod = 2000;       // Update data every 2 seconds
+unsigned long msgPeriod = 2000;       // Update data every 2 seconds
 
 // Callback function for MQTT message reception
 // It gets called every time there is an incoming message 
@@ -91,7 +91,7 @@ void callback(char* topic, byte* payload, unsigned int length){
   Serial.print("Message received [");
   Serial.print(topic);
   Serial.print("]: ");
-  for (int i = 0; i < length; i++) {
+  for (unsigned int i = 0; i < length; i++) {
     Serial.print((char)payload[i]);
   }
   Serial.println();
